Improve Layout component
